+++ src/DataTransferObjects/FakerMapData.php
@@ -2,12 +2,8 @@
 
 namespace Anteris\FakerMap\DataTransferObjects;
 
-use Faker\Generator;
-
 class FakerMapData
 {
-    protected Generator $faker;
-
     /** @var string The text we searched faker for. */
     public string $search;
 
@@ -21,24 +17,14 @@ class FakerMapData
     public ?int $letterDisparity;
 
     public function __construct(
-        Generator $faker,
         string $search,
         FakerMethod $method,
         ?float $percentageSimilar = null,
         ?int $letterDisparity = null
     ) {
-        $this->faker             = $faker;
         $this->search            = $search;
         $this->method            = $method;
         $this->percentageSimilar = $percentageSimilar;
         $this->letterDisparity   = $letterDisparity;
-    }
-
-    public function fake(...$customizations)
-    {
-        return call_user_func_array(
-            [$this->faker, $this->method->name],
-            $customizations ?? $this->method->parameters
-        );
     }
 }