+++ .github/CHANGELOG.md
@@ -4,9 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.0.0] - 2020-12-12
+
+### Added
+- `FakerResolver` class which will now make resolutions to Faker.
+- Data types to the dictionary.
+- Social security updates to dictionary.
+- Pass through to faker so the `FakerMap` class can be used just like faker as well.
+
+### Changed
+- API used for generating values to be more consistent and allow fallback to default values.
+
 ## [v0.1.0] - 2020-12-04
 ### Added
 - Initial release
 
+[v1.0.0]: https://github.com/Anteris-Dev/faker-map/compare/v0.1.0...v1.0.0
 [v0.1.0]: https://github.com/Anteris-Dev/faker-map/releases/tag/v0.1.0